Update README.md
In this program, we change the password of a user and check for errors using the require(), assert(), and revert() functions. The contract is named ChangePass, and it includes the user's name, old password, user address, and number. We then create a function named information in which we check the user's age to verify whether it is the user or someone else. This function returns the address of the person. Next, we create another function named change that changes the password from the old one to a new one. If the user does not perform this operation correctly, the password will not be changed, and the address of the person attempting to change the password will be returned.